Switch number to range input
Setting a number input on the iPad is annoying since it requires trying
to click in the right spot on the text box, backspacing, and then typing
the new number. A range is still annoying if you want to be precise, but
for coarse grained control it is much easier.

P.s.: also tried the inputmode="numeric" and pattern="[0-9]*" options on
the numeric input mentioned online, but they didn't seem to have any
effect on the iPad. Likely because the iPad is too old.